[feature] 구글로그인, token 관련 data 모듈 분리
