+++ .idea/gradle.xml
@@ -19,7 +19,9 @@
             <option value="$PROJECT_DIR$/data" />
             <option value="$PROJECT_DIR$/data/appinfo" />
             <option value="$PROJECT_DIR$/data/book" />
+            <option value="$PROJECT_DIR$/data/google_auth" />
             <option value="$PROJECT_DIR$/data/oneline" />
+            <option value="$PROJECT_DIR$/data/token" />
             <option value="$PROJECT_DIR$/data/user" />
             <option value="$PROJECT_DIR$/domain" />
             <option value="$PROJECT_DIR$/domain/appinfo" />
