+++ app/src/main/java/com/bookmark/bookmark_oneday/app/di/RepositoryModule.kt
@@ -2,8 +2,8 @@ package com.bookmark.bookmark_oneday.app.di
 
 import com.bookmark.bookmark_oneday.data.appinfo.repository.ApplicationRepositoryImpl
 import com.bookmark.bookmark_oneday.data.book.repository.SearchBookDetailRepositoryImpl
+import com.bookmark.bookmark_oneday.data.google_auth.repository.GoogleLoginRepositoryImpl
 import com.bookmark.bookmark_oneday.data.oneline.repository.OnelineRepositoryImpl
-import com.bookmark.bookmark_oneday.data.repository_impl.*
 import com.bookmark.bookmark_oneday.domain.appinfo.repository.ApplicationRepository
 import com.bookmark.bookmark_oneday.domain.book.repository.SearchBookDetailRepository
 import com.bookmark.bookmark_oneday.domain.login.repository.GoogleLoginRepository
