+++ data/google_auth/build.gradle
@@ -1,0 +1,48 @@
+plugins {
+    id 'com.android.library'
+    id 'org.jetbrains.kotlin.android'
+}
+
+android {
+    namespace 'com.bookmark.bookmark_oneday.data.google_auth'
+    compileSdk 33
+
+    defaultConfig {
+        minSdk 24
+        targetSdk 33
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles "consumer-rules.pro"
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+    kotlinOptions {
+        jvmTarget = '1.8'
+    }
+}
+
+dependencies {
+
+    implementation project(":core:api")
+    implementation project(":core:model")
+    implementation project(":domain:login")
+
+    implementation libs.androidx.core
+    implementation libs.androidx.appcompat
+    testImplementation libs.junit
+    androidTestImplementation libs.extJunit
+
+    // retrofit2 : https://square.github.io/retrofit/
+    implementation libs.bundles.retrofit
+
+    implementation libs.javaxInject
+}
