+++ data/google_auth/src/main/java/com/bookmark/bookmark_oneday/data/google_auth/model/response/GoogleReIssueAccessTokenResponse.kt
@@ -1,21 +1,20 @@
-package com.bookmark.bookmark_oneday.data.models.response_body
+package com.bookmark.bookmark_oneday.data.google_auth.model.response
 
 import com.bookmark.bookmark_oneday.domain.login.model.TokenInfo
 
-data class GoogleAccessTokenResponse(
+data class GoogleReIssueAccessTokenResponse(
     val access_token : String,
     val expires_in : Int,
     val scope : String,
     val token_type : String,
-    val refresh_token : String,
     val id_token : String?
 ) {
     companion object {
-        fun toTokenInfo(dto : GoogleAccessTokenResponse) : TokenInfo {
+        fun toTokenInfo(dto : GoogleReIssueAccessTokenResponse) : TokenInfo {
             return TokenInfo(
                 accessToken = dto.access_token,
                 expiredTime = dto.expires_in,
-                refreshToken = dto.refresh_token
+                refreshToken = null
             )
         }
     }
