+++ settings.gradle
@@ -27,3 +27,5 @@ include ':domain:login'
 include ':domain:appinfo'
 include ':data:oneline'
 include ':data:appinfo'
+include ':data:google_auth'
+include ':data:token'
